Revert "Bias the MEMC moves"
This reverts commit 3e385331aec13a20677f9ef93a99bb7bb6bf8507.

+++ src/ConfigSetup.h
@@ -243,7 +243,7 @@ struct MEMCVal {
 
 struct CFCMCVal {
   bool enable, readWindow, readRelaxSteps, readHistFlatness;
-  int window, relaxSteps;
+  long window, relaxSteps;
   double histFlatness;
   CFCMCVal(void) {
     readWindow = readRelaxSteps = readHistFlatness = false;
